More refactoring
The previous layout was already starting to creek and
felt like it wasn't going to scale well.

This layout feels better and will hopefullt be easier to
test if i ever get that far. Going forward im going to try
and keep as little code as possible in the cobra commands..